+++ ChangeLog.md
@@ -13,6 +13,7 @@
     - Have a guard against having both autoUsers=true and pwix:accounts-manager package
     - Remove assert() tests, replacing with logger.error() for debugging facility
     - Introduce email address and username cardinalities
+    - 'pwix_accounts_hub_list_all' now published also full roles of the target users
 
 ### 1.3.0